+++ tests/unit/stigviewer/main/import-ckl.tests.ps1
@@ -55,4 +55,3 @@ Describe "Import-CKL PS: $PSVersion" {
 
     }
 }
-